:sparkles: Start saving and loading folders locally
This isn't the extent of it yet, but it tests out the core idea. I'll expand
this to take in all the data and move to load-locally-first; then loading
from ToR becomes a refresh operation.

Ideally, eventually, maybe, I'll make this sort it works offline if ToR
isn't there for now.