:lipstick: Don't show a blog's name in list if we're just viewing it
That is: if we're viewing a specific subscription then don't include the
subscription title in every article listed in the article list, only show
the author name. On the other hand if we're viewing a folder then do show it
so we can know what subscription each article comes from (as well as the
author).